// src/modules/kubernetes/schemas/kubernetes-namespace.enum.ts
/**
 * Represents all Kubernetes namespaces in the Kubernetes cluster
 */
export enum KubernetesNamespace {
  AgoraCloudProduction = 'agoracloud-prod',
  AgoraCloudDevelopment = 'agoracloud-dev',
  AgoraCloudSaid = 'agoracloud-said',
  AgoraCloudWaleed = 'agoracloud-waleed',
  AgoraCloudMarc = 'agoracloud-marc',
}